extracted grouping unifying code to a dedicated file
